Significantly refactor the various model classses to experiment with new ways of organizing the key neural network layers
