+++ testing/unit_tests/test_model.py
@@ -1,3 +1,7 @@
+import ivadomed.architecture.blocks
+import ivadomed.architecture.densenet
+import ivadomed.architecture.resnet
+import ivadomed.architecture.unet
 import ivadomed.models as imed_model
 import torch
 import torchvision
@@ -15,7 +19,7 @@ def test_model_3d_att():
     # verifying if 3d attention model can be created
     a = [[[[[0 for i in range(48)] for j in range(48)] for k in range(16)]]]
     inp = torch.tensor(a).float()
-    model = imed_model.Modified3DUNet(in_channel=1, out_channel=1, attention=True)
+    model = ivadomed.architecture.unet.Modified3DUNet(in_channel=1, out_channel=1, attention=True)
     inf = model(inp)
     assert(type(inf) == torch.Tensor)
 
@@ -23,7 +27,7 @@ def test_model_3d_att():
 def test_resnet():
     a = [[[[0 for i in range(100)] for i in range(100)]]]
     inp = torch.tensor(a).float()
-    model = imed_model.ResNet(torchvision.models.resnet.BasicBlock, [2, 2, 2, 2])
+    model = ivadomed.architecture.resnet.ResNet(torchvision.models.resnet.BasicBlock, [2, 2, 2, 2])
     inf = model(inp)
     assert (type(inf) == torch.Tensor)
 
@@ -31,7 +35,7 @@ def test_resnet():
 def test_densenet():
     a = [[[[0 for i in range(100)] for i in range(100)]]]
     inp = torch.tensor(a).float()
-    model = imed_model.DenseNet(32, (6, 12, 24, 16), 64)
+    model = ivadomed.architecture.densenet.DenseNet(32, (6, 12, 24, 16), 64)
     inf = model(inp)
     assert (type(inf) == torch.Tensor)
 
@@ -39,7 +43,7 @@ def test_densenet():
 def test_filmed_unet():
     a = [[[[0 for i in range(100)] for i in range(100)]]]
     inp = torch.tensor(a).float()
-    model = imed_model.FiLMedUnet()
+    model = ivadomed.architecture.unet.FiLMedUnet()
     inf = model(inp)
     assert (type(inf) == torch.Tensor)
 
@@ -47,7 +51,7 @@ def test_filmed_unet():
 def test_film_generator():
     a = [[[[0 for i in range(64)] for i in range(64)]]]
     inp = torch.tensor(a).float()
-    model = imed_model.FiLMgenerator(64, 1)
+    model = ivadomed.architecture.blocks.FiLMgenerator(64, 1)
     inf = model(inp)
     assert (type(inf[0]) == torch.Tensor)
     assert (type(inf[1]) == list)
