+++ testing/unit_tests/test_onnx.py
@@ -3,6 +3,8 @@ import torch
 import numpy as np
 import shutil
 import logging
+
+import ivadomed.architecture.unet
 from ivadomed import utils as imed_utils
 from ivadomed import inference as imed_inference
 from ivadomed import models as imed_models
@@ -24,7 +26,7 @@ LENGTH_3D = (112, 112, 112)
 
 
 def test_onnx(download_data_testing_test_files):
-    model = imed_models.Modified3DUNet(1, 1)
+    model = ivadomed.architecture.unet.Modified3DUNet(1, 1)
     if not PATH_MODEL.exists():
         PATH_MODEL.mkdir()
     torch.save(model, PATH_MODEL_PT)
